executeCommand()-functionality and test
Added functionality to the library:
You may now call the executeCommand(command, callbackFn) - function in
order to execute some commandline-command on the current machine.

Also added a test which will show a directory listing if running
correctly. This does support linux as well as windows and mac, depending
on the commands you are running.